\name{cluster_members}
\alias{cluster_members}
\title{Return indices of BGUs belonging to a given cluster}
\usage{
cluster_members(cluster_id, sol)
}
\arguments{
\item{cluster_id}{Integer, cluster label.}
\item{sol}{Integer vector of assignments.}
}
\value{
Integer vector of indices.
}
\description{
See R/01_utils.R for details.
}
