\name{fix_misallocations}
\alias{fix_misallocations}
\title{fix_misallocations}
\usage{
fix_misallocations(best_sol, W, adj, row_W, col_W, Pop_min = 500, Pop_max = 3500)
}
\arguments{
\item{best_sol}{See Description.}
\item{W}{See Description.}
\item{adj}{See Description.}
\item{row_W}{See Description.}
\item{col_W}{See Description.}
\item{Pop_min}{See Description.}
\item{Pop_max}{See Description.}
}
\value{
See Description.
}
\description{
See R/07_adsa_almd.R for details.
}
