\name{operator_11}
\alias{operator_11}
\title{operator_11}
\usage{
operator_11(sol, W, adj, row_W, col_W, Pop_max = Inf, ...)
}
\arguments{
\item{sol}{See Description.}
\item{W}{See Description.}
\item{adj}{See Description.}
\item{row_W}{See Description.}
\item{col_W}{See Description.}
\item{Pop_max}{See Description.}
\item{...}{Additional arguments (see Description / source).}
}
\value{
See Description.
}
\description{
See R/06_operators.R for details.
}
