\name{operator_2}
\alias{operator_2}
\title{operator_2}
\usage{
operator_2(sol, W, adj, row_W, col_W, ...)
}
\arguments{
\item{sol}{See Description.}
\item{W}{See Description.}
\item{adj}{See Description.}
\item{row_W}{See Description.}
\item{col_W}{See Description.}
\item{...}{Additional arguments (see Description / source).}
}
\value{
See Description.
}
\description{
Operator 2: Expand smallest cluster by absorbing one border BGU
}
