\name{operator_4}
\alias{operator_4}
\title{operator_4}
\usage{
operator_4(sol, W, adj, row_W, col_W, ...)
}
\arguments{
\item{sol}{See Description.}
\item{W}{See Description.}
\item{adj}{See Description.}
\item{row_W}{See Description.}
\item{col_W}{See Description.}
\item{...}{Additional arguments (see Description / source).}
}
\value{
See Description.
}
\description{
Operator 4: Remove a border BGU from its cluster, assign to best neighbour
}
