\name{operator_5}
\alias{operator_5}
\title{operator_5}
\usage{
operator_5(sol, W, adj, row_W, col_W, ...)
}
\arguments{
\item{sol}{See Description.}
\item{W}{See Description.}
\item{adj}{See Description.}
\item{row_W}{See Description.}
\item{col_W}{See Description.}
\item{...}{Additional arguments (see Description / source).}
}
\value{
See Description.
}
\description{
Operator 5: Include a border BGU from an adjacent cluster
}
