\name{operator_9}
\alias{operator_9}
\title{operator_9}
\usage{
operator_9(sol, W, adj, row_W, col_W, ...)
}
\arguments{
\item{sol}{See Description.}
\item{W}{See Description.}
\item{adj}{See Description.}
\item{row_W}{See Description.}
\item{col_W}{See Description.}
\item{...}{Additional arguments (see Description / source).}
}
\value{
See Description.
}
\description{
Operator 9: Exchange border BGUs between two adjacent clusters
}
