\name{sara_iterated}
\alias{sara_iterated}
\title{sara_iterated}
\usage{
sara_iterated(W, adj, row_W, col_W, SC_min, Pop_min, Pop_max, max_it = 5, verbose = TRUE, temp = NULL)
}
\arguments{
\item{W}{See Description.}
\item{adj}{See Description.}
\item{row_W}{See Description.}
\item{col_W}{See Description.}
\item{SC_min}{See Description.}
\item{Pop_min}{See Description.}
\item{Pop_max}{See Description.}
\item{max_it}{See Description.}
\item{verbose}{See Description.}
\item{temp}{See Description.}
}
\value{
See Description.
}
\description{
See R/05_sara.R for details.
}
