\name{sara_max_pop_score}
\alias{sara_max_pop_score}
\title{sara_max_pop_score}
\usage{
sara_max_pop_score(sol, W, row_W, col_W, Pop_max)
}
\arguments{
\item{sol}{See Description.}
\item{W}{See Description.}
\item{row_W}{See Description.}
\item{col_W}{See Description.}
\item{Pop_max}{See Description.}
}
\value{
See Description.
}
\description{
See R/05_sara.R for details.
}
