\name{update_op_tracker}
\alias{update_op_tracker}
\title{update_op_tracker}
\usage{
update_op_tracker(tracker, op_idx, sol_before, sol_after, F_before, F_after, accepted, improved_best)
}
\arguments{
\item{tracker}{See Description.}
\item{op_idx}{See Description.}
\item{sol_before}{See Description.}
\item{sol_after}{See Description.}
\item{F_before}{See Description.}
\item{F_after}{See Description.}
\item{accepted}{See Description.}
\item{improved_best}{See Description.}
}
\value{
See Description.
}
\description{
See R/06_operators.R for details.
}
