\name{fix_contiguity}
\alias{fix_contiguity}
\title{fix_contiguity}
\usage{
fix_contiguity(sol, adj, max_iter = 5000)
}
\arguments{
\item{sol}{See Description.}
\item{adj}{See Description.}
\item{max_iter}{See Description.}
}
\value{
See Description.
}
\description{
See R/06_operators.R for details.
}
