\name{operator_1}
\alias{operator_1}
\title{operator_1}
\usage{
operator_1(sol, W, adj, row_W, col_W, ...)
}
\arguments{
\item{sol}{See Description.}
\item{W}{See Description.}
\item{adj}{See Description.}
\item{row_W}{See Description.}
\item{col_W}{See Description.}
\item{...}{Additional arguments (see Description / source).}
}
\value{
See Description.
}
\description{
Operator 1: Merge least-valid cluster into best neighbour
}
