\name{operator_10}
\alias{operator_10}
\title{operator_10}
\usage{
operator_10(sol, W, adj, row_W, col_W, best_sol = NULL, ...)
}
\arguments{
\item{sol}{See Description.}
\item{W}{See Description.}
\item{adj}{See Description.}
\item{row_W}{See Description.}
\item{col_W}{See Description.}
\item{best_sol}{See Description.}
\item{...}{Additional arguments (see Description / source).}
}
\value{
See Description.
}
\description{
See R/06_operators.R for details.
}
