\name{penalty_pop_max_total}
\alias{penalty_pop_max_total}
\title{penalty_pop_max_total}
\usage{
penalty_pop_max_total(pop_vec, Pop_max, r)
}
\arguments{
\item{pop_vec}{See Description.}
\item{Pop_max}{See Description.}
\item{r}{See Description.}
}
\value{
See Description.
}
\description{
See R/03_objective.R for details.
}
