\name{penalty_pop_min_single}
\alias{penalty_pop_min_single}
\title{penalty_pop_min_single}
\usage{
penalty_pop_min_single(pop, Pop_min, r)
}
\arguments{
\item{pop}{See Description.}
\item{Pop_min}{See Description.}
\item{r}{See Description.}
}
\value{
See Description.
}
\description{
See R/03_objective.R for details.
}
