\name{penalty_pop_min_total}
\alias{penalty_pop_min_total}
\title{Total population penalty   [Equation 14 summed]}
\usage{
penalty_pop_min_total(pop_vec, Pop_min, r = 3)
}
\arguments{
\item{pop_vec}{See Description.}
\item{Pop_min}{Numeric.}
\item{r}{Integer.}
}
\value{
Numeric scalar.
}
\description{
PPop(x) = sum_\{Ms in M\} PPop(Ms)
}
