\name{sha}
\alias{sha}
\title{sha}
\usage{
sha(bgus, sol, W, adj, row_W, col_W, SC_min, Pop_min, max_iter = 6000)
}
\arguments{
\item{bgus}{See Description.}
\item{sol}{See Description.}
\item{W}{See Description.}
\item{adj}{See Description.}
\item{row_W}{See Description.}
\item{col_W}{See Description.}
\item{SC_min}{See Description.}
\item{Pop_min}{See Description.}
\item{max_iter}{See Description.}
}
\value{
See Description.
}
\description{
See R/09_sha.R for details.
}
